Check version existence asynchronously.
`gulp-ssh` uses under the hood `ssh2-streams`. Due to a bug in
`ssh2-streams` > 0.0.14, every time we were pushing to CDN we were
receiving the error `Error: no writecb in Transform class`. If inside
the gulp task we would have returned the stream (`return gulp.src....`)
the error was still happening but gulp was catching it. It didn't appear
in CLI leaving the impression that everything was ok.

More details here: https://github.com/teambition/gulp-ssh/issues/32.
I am forcing the use of `ssh2-streams v0.0.14` for the moment.